Refactoring and additional modes (#53)
Adding the winning states to the solutions organizer as they are created
instead of pushing them into the heap is a small win on its own; we
might need to start evicting from the solutions! This unfortunately
makes the rescaling a little less likely to occur, but we could change
that to focus on unique solutions in particular.

The other modes are performed infrequently in additional ways to try
focusing further on various states. Tweaks will be necessary.