Add new search modes PickMin and Dependent.
Assign modes to threads by index. Dependent is the mode
that will pick another mode based on other factors.
PickMin picks based on min elapsed time. I'd prefer this be assigned to
a thread rather than part of Dependent (which is essentially random),
but it locks the queue for a long time, so for now it's set to be rare.