Implement graph condensing. (#72)
Adds some additional world graph features: same_area, same_graph, and
best_movements.
Adds an "always" feature for Exits with no price and no access rule.

This generates CondensedEdges for every spot, since we don't bother to
check that exits and actions don't move you to spots of interest only.
For AV2, the graph is reduced from 465 spots to 249 spots of interest
with 355 edges between them (and 437 sources with 709 edges total).

CondensedEdges are not yet in use.