Compare to the best elapsed time before skipping.
We were skipping the update of a state record if its current time
was more than the best time; since we started removing the loopback,
it was possible for a record that subtracted the loopback to beat
an actual time that hadn't yet subtracted it. So the solution is to
subtract it before making that comparison.

This fixes #105.